thay đổi phân quyền trong config
